backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
et-xmlfile==1.1.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
numpy==1.21.6
openpyxl==3.1.5
pandas==1.5.3
pip==24.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
